TCP profile: always transmit ACK number even if ACK flag = 0
This change almost completely reverts change at revision 1086 entitled:
  TCP profile: do not transmit ACK number if ACK flag = 0 and ACK number = 0

The revert is required because this optimisation cannot handle correctly the
case where N subsequent TCP packets are sent with ACK flag = 0 and the very
same non-zero ACK number. If N is large enough, the variable_length_32_enc()
encoding method will need 0 bit to be transmitted (static behavior), but the
decompressor will mistaken it as ACK number = 0. Optimisation is not possible
in all cases, so remove it.