decompression: correctly handle padding bytes
Padding bytes are located at the very beginning of the ROHC packet if there
is any of them. They must be followed by feedback data and/or a ROHC header.
One padding-only ROHC packet is not valid.

Add a new test for padded ROHC packets:
 * 1-byte padding + IR
 * 1-byte padding + feedback
 * 2-byte padding + feedback
 * 1-byte padding + feedback + IR
 * 5-byte padding
The first 4 use cases shall succeed, the last one shall fail.