Enhance the ROHC API to better handle received feedback information
Here is a short explanation of feedback management in the ROHC framework
(see RFC 3095 for more details). Let's try to transmit compressed information
from node n1 to node n2.

Each node got a compressor and a decompressor. Compressor A compresses data,
sends it to decompressor A (1) that then decompresses it. Compressor A and
decompressor A make a pair. If decompressor A wants to send some feedback
information to compressor A, it asks the same-side associated compressor B
to transmit it (2). Compressor B transmits the feedback information to
decompressor B (3). When decompressor B receives the feedback, it gives that
information to the same-side associated compressor A (4).

          node n1                                 node n2
 +---------------------+                 +---------------------+
 |                     |                 |                     |
 | +----------------+  |                 | +----------------+  |
 | |  compressor A  |--------(1)---------->| decompressor A |  |
 | +----------------+  |                 | +----------------+  |
 |         ^           |                 |         | (2)       |
 |         | (4)       |                 |        \/           |
 | +----------------+  |                 | +----------------+  |
 | | decompressor B |<-------(3)---------->|  compressor B  |  |
 | +----------------+  |                 | +----------------+  |
 |                     |                 |                     |
 +---------------------+                 +---------------------+

The change in source code is related to steps (3) and (4). When decompressor B
decompresses the feedback information emitted by compressor B on behalf of
decompressor A, decompressor B used to automatically gives the feedback information
to compressor A in a transparent way (ie. no user interaction). Decompressor B
used to call rohc_comp_deliver_feedback(compressor A, feedback) directly.

This used to have a major drawback: the compression (compressor A) and
decompression (decompressor B) actions on one single node couldn't be run in
separate threads. The decompression action could modify the compression context
with no synchronisation mechanism at all, so thread cannot be used efficiently.

Now, at step (4), the feedback information transmitted by decompressor A to its
remote peer compressor A can now be retrieved thanks to an additional parameter
added to the rohc_decompress3() function:

	int rohc_decompress3(struct rohc_decomp *const decomp,
	                     const struct rohc_buf rohc_packet,
	                     struct rohc_buf *const uncomp_packet,
	                     struct rohc_buf *const rcvd_feedback);

By removing the direct call, the user gets in charge of retrieving the received
feedback during decompression and delivering it to the related compressor. The
user gains control on the feedback transmission and can use locks if needed.

To achieve this, feedback-related code was reworked:
 * rohc_comp_deliver_feedback2() is introduced to accept a struct rohc_buf as
   parameter and accept several feedback items at once.
 * code for parsing feedback is now shared between compression and decompression
   libraries (it is located in common, files feedback_parse.c/h).
 * code for creating feedback was moved from src/decomp/feedback.c/h to
   src/decomp/feedback_create.c/h
 * ROHC_FEEDBACK_ONLY return code is not returned any more by rohc_decompress3()
	because checking whether rcvd_feedback is empty gives the same information.
 * ROHC_NON_FINAL_SEGMENT return code is not returned any more by rohc_decompress3()
	because we can now check if uncomp_packet is empty or not, this gives the same
	information.