Enhance the ROHC API to better handle feedback piggybacking
Here is a short explanation of feedback management in the ROHC framework
(see RFC 3095 for more details). Let's try to transmit compressed information
from node n1 to node n2.

Each node got a compressor and a decompressor. Compressor A compresses data,
sends it to decompressor A (1) that then decompresses it. Compressor A and
decompressor A make a pair. If decompressor A wants to send some feedback
information to compressor A, it asks the same-side associated compressor B
to transmit it (2). Compressor B transmits the feedback information to
decompressor B (3). When decompressor B receives the feedback, it gives that
information to the same-side associated compressor A (4).

          node n1                                 node n2
 +---------------------+                 +---------------------+
 |                     |                 |                     |
 | +----------------+  |                 | +----------------+  |
 | |  compressor A  |--------(1)---------->| decompressor A |  |
 | +----------------+  |                 | +----------------+  |
 |         ^           |                 |         | (2)       |
 |         | (4)       |                 |        \/           |
 | +----------------+  |                 | +----------------+  |
 | | decompressor B |<-------(3)-----------|  compressor B  |  |
 | +----------------+  |                 | +----------------+  |
 |                     |                 |                     |
 +---------------------+                 +---------------------+

The change in source code is related to steps (2) and (3). When decompressor A
decompresses the ROHC packet emitted by compressor A, decompressor A might want
to transmit some feedback information to compressor A. The decompressor A used
to automatically gives the feedback information to compressor B in a transparent
way (ie. no user interaction). Indeed, the decompressor A used to call the function
rohc_comp_piggyback_feedback(compressor A, feedback_data, feedback_size) directly.

This used to have a major drawback: the decompression (decompressor A) and
compression (compressor B) actions on one single node (node B) couldn't be run
in separate threads. The decompression action could modify the compression context
with no synchronisation mechanism at all, so thread cannot be used efficiently.

Now, at step (2), the feedback information generated by decompressor A for its
remote peer compressor A can now be retrieved thanks to an additional parameter
added to the rohc_decompress3() function:

	int rohc_decompress3(struct rohc_decomp *const decomp,
	                     const struct rohc_buf rohc_packet,
	                     struct rohc_buf *const uncomp_packet,
	                     struct rohc_buf *const rcvd_feedback,
	                     struct rohc_buf *const feedback_send);

By removing the direct call, the user gets in charge of retrieving the feedback
to be sent during decompression and sending it on the remote peer for the
compressor A. The user may perform this action in two different ways: he/she
sends the data on his/her own (it entirely depends on the user application) or
he/she uses piggybacking by prepending the content of `feedback_send` before any
ROHC packet generated by the compressor B. The decompressor B will retrieve the
feedback data and gives it to the user through the `rcvd_feedback` parameter of
the rohc_decompress3() function. The user gains control on the feedback
transmission and can use locks if needed.

Note that the old API still works the way it always worked. The rohc_decompress(),
rohc_decompress2() and rohc_decompress_both() functions automatically sends
feedback information to the same-side associated compressor. And, the rohc_compress(),
rohc_compress2() and rohc_compress3() functions still piggyback that feedback
information automatically.