Merge the 3 librohc_*.so libraries into one single librohc.so library.
The 3 libraries are used together almost everytime, so better have one single
library. It avoids exposing some private parts of the librohc_common.so
library.

Also remove the dllexport.h and its ROHC_EXPORT constant, and the symbols of
the now-internal librohc_*.so libraries. Also, rename of the source files
that shared the same name in compression and decompression parts: includes
are now clearer.