Update license and copyright of src/common/protocols/udp.h
The src/common/protocols/udp.h file mimics the udphdr struct defined by
netinet/udp.h from the GNU C library. It is defined in that file to be portable
on all platforms, even the platforms that miss the declarations or got different
declarations, such as Microsoft Windows or FreeBSD. The udphdr struct being
trivial, the original copyrights and licenses are not kept to simplify the
library license.