W-LSB values shall be acknowledged only in R-mode
According to RFC 3095, §4.5.2, W-LSB values shall be acknowledged only in
R-mode. In U/O-mode, a sliding window with a limited maximum width is used.

This change reduce a little bit the library performances, but it ensures better
conformance with RFC and it ensures that compression always works: a TCP stream
was added as non-regression test.