Don't use LateX syntax in Doxygen comments.
LateX syntax is only used to pretty display the formulas of the f() and g()
functions defined by RFC 3095. The formulas are quite simple, so pretty display
is not really needed. This changes will allow us to remove the Doxygen
dependency on LateX, thus the ROHC library dependency on LateX.