Remove the Doxygen dependency on LateX.
LateX syntax was only used to pretty display the formulas of the f() and g()
functions defined by RFC 3095. The formulas are quite simple, so pretty display
was not really needed, and was removed with previous change. Let's tell autoconf
to not check for LateX binary anymore.

This changes will allow the ROHC library to not dependent on LateX any more.
Distribution packagers should appreciate :)