Factorize the decompression part of the RFC3095-based and TCP profiles
The general algorithm to decompress a ROHC packet can be shared be the
RFC3095-based profiles (IP-only, UDP, UDP-Lite, RTP, ESP) and the TCP
profile:
 a. parse the ROHC header
 b. for IR and IR-DYN packet, check for correct compressed header (CRC)
 c. decode extracted bits
 d. build uncompressed headers (and check for correct decompression
    for non-IR(-DYN) packets)
 e. copy the payload (if any)
 f. update the compression context

Steps c and d may be repeated twice if packet or context repair is
attempted upon the detection of a CRC failure.

This change avoids the duplication of the above algorithm between the
common code of the RFC3095-based profiles and the TCP profile.