Fix bug #1407084: TCP profile: DSCP not correctly handled https://bugs.launchpad.net/rohc/+bug/1407084
The ROHC library didn't handle TCP streams with non-zero DSCP fields. The
ROHC compressor built the co_common packet in a wrong way. The DSCP field
is not sent in the 6 LSB bits of the 8-bit field, but in the 6 MSB bits.
This leads to the ROHC packet being malformed and the ROHC decompressor
to mis-decompress it. The CRC check shall detect the problem, so it
leads to a packet loss.

The problem was found with a SSH network stream that got the DSCP field
set to a non-zero value. The stream was captured on my home LAN with the
ROHC sniffer, and then anonymized before its addition to the
non-regression tests.