Fix bug #1407271: TCP profile: decompressor fails to decode duplicate SYN/ACK
The ROHC library didn't handle TCP streams with duplicate SYN/ACK.
The ROHC decompressor parsed the IP dynamic part of the IR packet
in a wrong way (forget to skip some bits after parsing them). This
leads to the ROHC packet being considered as malformed and dropped.

The problem was found with a HTTP network stream that was captured
on my home LAN with the ROHC sniffer, and then anonymized before
its addition to the non-regression tests.

See https://bugs.launchpad.net/rohc/rohc-1.7.x/+bug/1407271