API: introduce ROHC_PACKET_MAX as the max number of ROHC packet types
Up to now, there was no proper way to iterate over all the types of ROHC
packets. The ROHC sniffer hardcoded the packet type with the larger
identifier, it works but it is fragile if a new packet type is added.

So, introduce ROHC_PACKET_MAX (with no numeric value set) for that use.
Update the ROHC sniffer to make use of it.