TCP profile: do not assert because of malformed MSS option
The AFL fuzzer detected a problem with a malformed MSS option. If one
TCP packet with 20 bytes of options got as first option one MSS option
that wrongly reports to be 20-byte long, other options are then ignored
during the detection of option list changes. But they are not by the
function that builds the TCP part of the dynamic chain (well-known
length of the MSS option is used instead). When handling the second
option, the library asserts because no index was assigned to it.

Solution: check if well-known options with well-known lengths are
well-formed during the building of the dynamic chain, but also while
checking if the TCP profile matches the packet. The first checks are
for robustness, the second ones are to force the packet to use the
IP-only profile instead of the IP/TCP profile when malformed options
are detected.