TCP profile: be stricter when accepting IPv4 packets
Make the TCP profile rejects IPv4 packets that it does not support:
* IPv4 packets that are fragments
* IPv4 packets with IP options

Make the TCP profile rejects the IPv4 packets with bad IP checksums. The
TCP profile does compress/decompress them without major problems but the
decompressed packet won't match the source packet because the IP
checksum is not transmitted but computed (correctly) during the
decompression.