TCP profile: do not accept packets with non-zero option padding
The TCP protocol and the ROHC IP/TCP profile expect that the TCP options
are padded with zero bytes if they are not aligned on 4 bytes. While the
ROHC library compresses/decompresses successfully TCP packets with
non-zero padding bytes, the decompressed packet does not match the
original packet since padding bytes are not transmitted but generated
by the decompressor.

Making the TCP profile rejects TCP packets with unexpected padding bytes
allows the ROHC library to fallback on the IP-only profile for the
decompression.