TCP profile: recycle indexes for lists of TCP options when needed
TCP options are transmitted with compressed lists. The options are
identified by an index. Some well-known TCP options got a fixed index
value. Other options don't. The ROHC compressor shall then find an index
to compress them.

Up to now, the compressor was searching for an index that was never used
before by the compressor context. This doesn't scale if the TCP stream
contains more than 16 different options. An index shall be re-used.

This change adds a simple algorithm to recycle the oldest index. This
should avoid re-using the index of an heavily used option.