Enfore rules from RFC2460 about IPv6 extension headers
The RFC 2460 that defines the IPv6 protocol and the IPv6 extension
headers defines the following rules about the usage of extension
headers:

§4 reads:
  The Hop-by-Hop Options header, when present, must
  immediately follow the IPv6 header.

§4.1 reads:
  Each extension header should occur at most once, except for the
  Destination Options header which should occur at most twice (once
  before a Routing header and once before the upper-layer header).

The ROHC library (both RFC3095-based profiles and the TCP profile)
did not enforce those rules. However, not doing so makes impossible
for the library to correctly compress IPv6 extension headers.

If 2 extension headers of the same type are present with different
contents, they will share the same list index. So, decompression
will fail to generate the very same content as the input packet for
both extension headers.

This change adds the 2 checks for both RFC3095-based and TCP profiles.