TCP profile: handle gracefully MSS and WS options in the irregular chain
The MSS and WS options should only be present on SYN packets, so they
shall not be transmitted in the irregular chain. However, malformed
packets might do so.

This changes ensures that MSS and WS options get correctly decompressed
when transmitted in the irregular chain. The MSS and WS payloads are
retrieved from the context since the irregular encoding of MSS and WS
options is empty.