TCP profile: handle EOL option with changing sizes
The EOL option is 1-byte long. However, the ROHC protocol compresses the
following zero bytes of padding along with the EOL option. So, the EOL
option might by 1-byte to 4-byte long.

The 1/4-byte EOL option was already handled by the TCP profile of the
library. However, compression was wrong sometimes: the EOL option item
shall be transmitted if the length of the EOL option change.