RFC 3095 profiles: avoid crash on empty inner IPv4 header
The compression profiles based on RFC 3095 handles up to two IP headers,
either IPv4 or IPv6. If the packet is malformed in a way that makes the
inner IP header empty, the library crashes.

If the inner IP header got an unknown version number or is malformed,
the code that searches for the best compression profile tries to print
a helpful trace to the library users. When doing so, it uses the very
first byte of the inner IP header without checking first if that byte
is available. This change adds such a check.