TCP profile: handle TCP stream with zero URG flag but changing URG pointer
The URG pointer is not correctly compressed if the URG flag is not set
but the URG pointer is not zero.

The ROHC protocol aims to transmit all the fields unchanged even if
their combination doesn't make sense. Here the non-zero URG pointer
is useless with the zero URG flag, but the ROHC compressor shall
transmit it so that the decompressor is able to decompress the exact
same packet, byte per byte.