TCP profile: handle non-zero static URG pointer
If the value of the URG pointer is not zero and does not change across
several packets, the compressor may transmit zero bit of it in the last
packets. If the compressor does so, the decompressor shall take the
value from its content.

This change records the URG pointer in the context and uses it to build
the uncompressed packet if no bit of URG pointer is transmitted by the
compressor.