TCP profile: handle changing TCP MSS and WS options
The TCP Maximum Segment Size (MSS) option and the Window Scale (WS)
option should not change of value for the lifetime of the TCP
connection. However the MSS and WS options could change if the
compression context is re-used for another TCP connection. So, changing
TCP MSS and/or WS options shall be handled correctly by the ROHC
compressor.