TCP profile: handle not-always-present TCP EOL option
The TCP EOL option is compressed with its length. The following scenario
was not correctly handled:
* packet  N-1: the EOL option is present with length L,
* packets N-M: the EOL option is not present at all,
* packet  M+1: the EOL option is present with length L.

Packet M+1 was not correctly compressed/decompressed. The decompressed
EOL option was 0-byte length.

This was due to packet M+1 not sending any item for the EOL option in
the list of options after the co_common body. The item for the EOL
option was sent in the irregular chain, but as an empty item since the
irregular chain cannot change the EOL length. The decompressor then took
the default value of zero as EOL length instead of using the one sent in
packet N-1.

The change adds a flag that is set during the parsing of dynamic/irregular
option list. This flags then allows the decompressor to use the context
value instead of the packet value if the item for EOL in the dynamic
chain is absent and the one in the irregular chain is empty.