RFC3095 profiles: context repair: use LSB ref-1 for IP-ID offset
In case of CRC failure, context repair may be attempted by the
decompressor. In such a case, packet decompression is attempted with
ref -1 for the W-LSB decoding of SN instead of ref 0.

Although the RFC 3095 only describes this mechanism for SN, performing
the same for IP-ID offset seems obvious. This patch does that. It helps
in correcting the context more quickly.