Handle corner cases of W-LSB encoding with variable p parameter
The RTP SN, ESP SN and RTP TS are encoded with the W-LSB scheme. The
shift parameter p is variable:
 a/ below of specified number of bits, it defaults to a static value,
 b/ above that specified number of bits, it is computed from that number
   of bits.

This behavior of the RTP SN, ESP SN and RTP TS may lead the compressor
to create ROHC packets that the decompressor cannot decode successfully.

The problem happens if the conditions hereafter are fulfilled:
 * choosing a packet that uses case a/ is possible
 * choosing a packet that uses case b/ is possible but some of them
   cannot be chosen
 * the number of bits required for W-LSB encoding is stricly smaller for
   case a/ than for case b/
 * another packet field cannot be heavily compressed and a larger packet
   type is required (eg. UOR-2* with extension 3 instead of UO-0).

In such a case, more bits of SN/TS than expected are transmitted. The
number of bits computed for case a/ is used to choose a packet type
that should in fact use case b/ instead. The robustness of the field
may be therefore lowered, and the decompressor may fail to decode it
correctly. It may happen if there is a little bit of packet reordering
between the compressor and decompressor, eg. SN/TS goes backward a
little bit.

The change ensures that the packet types using case b/ are only chosen
if case b/ is possible for them, and that the packet types using case a/
are only chosen if case a/ is possible for them.

The problem was discovered with AFL. AFL generated an ESP stream with
moderate packet reordering that failed to be decompressed. The ESP
stream was added as a non-regression test.