TCP profile: fix the way TCP options indexes are recycled
The TCP profiles splits TCP options into 2 groups:
 * common options such as EOL, NOP, MSS, WS...
 * less common options or unknown options

The options of the first group are assigned static well-known indexes in
the range [0-6]. The options of the second group are assigned indexes
from the range [7-15] dynamically.

Whenever a new option of the 2nd group is encountered in a TCP stream, a
new free index is reserved in the range [7-15]. When no free index is
available any more, the oldest index in range [7-15] is recycled for the
new option. This mechanism was added by change 869bbeab a few months ago.

If two new options are encountered at the same time in one TCP packet of
a TCP stream and all indexes in range [7-15] are already reserved, then
the two oldest indexes shall be recycled.

Such a case was not correctly handled: the oldest index was recycled for
the first option, then re-recycled for the 2nd option. The recycled
index was not reset as it should be.

The problem was detected with AFL. The generated network capture was
added as a new non-regression test.