TCP profile: fix the computation of the TCP payload
The TCP payload is the whole packet minus all of the IP headers
(including their extension headers if any), the TCP header, and
all of the TCP options.

The code was wrong and the IP headers and their extension headers
were not taken into account. The only side effect detected so far
was the inappropriate use of the scaled encoding of the TCP
sequence number and the use of packet types that shouldn't be used
when the TCP payload is empty.

The problem was detected with the ROHC sniffer on my personal LAN.