TCP profile: fix (again) recyling of TCP options indexes
The TCP profiles splits TCP options into 2 groups:
 * common options such as EOL, NOP, MSS, WS...
 * less common options or unknown options

Previous changes 869bbeab and 4ae9c314 fixed the way the 2nd group of
options was handled:
 * Whenever a new option of the 2nd group is encountered in a TCP stream,
   a new free index is reserved in the range [7-15]. This part was
   present since ever.
 * When no free index is available any more, the oldest index in range
   [7-15] is recycled for the new option. This part was added with
   869bbeab.
 * If two new options are encountered at the same time in one TCP packet
   of a TCP stream and all indexes in range [7-15] are already reserved,
   then the two oldest indexes shall be recycled with 2 different
   indexes. This part was added with 4ae9c314.

Another problem was detected with AFL. If a new TCP packet contains 2
options (one option that was assigned an index very long ago and one
option that is brand new), then the ROHC compressor will:
 * find that the first option was previously assigned an index,
 * find the oldest index that may be recycled for the 2nd option.
If the 2 indexes may be the same if the first option was not used during
a long time. In such a case, the ROHC packet encodes the 2 options with
the very same index. The ROHC decompressor doesn't like that and rejects
the ROHC packet.

The network capture generated by AFL was added as a new non-regression
test.