TCP profile: properly compare and record changed options into context
The TCP profile compares the TCP options found in packet with the ones
found in the compression context. The TCP profile also record changed
options in the context.

The way both actions were done was messy. This change factorizes the
comparison and record actions into separate functions and makes use of
them to replace copied/pasted code.

This change detected a problem with the record of SACK options in the
context when they were transmitted in the dynamic chain. The compression
of the next packet therefore detected another superfluous option change.

This change was performed because AFL found a problem with the EOL
option that was similar to one encountered with the SACK option.

Several non-regression tests were updated to reflect the SACK change. A
new non-regression test was added for the EOL bug.