TCP profile: improve code readability of co_common packet building
The co_baseheader is the function that builds all the packet types that
are specific to TCP, ie. rnd_X, seq_X and co_common.

The code that builds all the rnd_X and seq_X packets was already split
into separate smaller functions. That was not the case of the code for
the co_common packet.

So, this change splits the code that builds the co_common packet type
into a separate smaller function.

This change makes the co_baseheader function leave the top list of
complexe function reported by GNU complexify. That's nice :)