RFC3095 list compression: handle anonymous lists
The ROHC decompressor was ready for anonymous compressed lists, but the
compressor didn't use them. This change allows the compressor to use
them.

The rational to choose an anonymous list is:
 * if no reference list was fully transmitted to the decompressor yet,
   do not use anonymous lists ;
 * otherwise, use anonymous lists until at least 2 packets transmit the
   same list, then switch to an identified list with a gen_id.

On network streams with lists that follow the pattern:
  L1 L1 ... L1 L1 L2 L2 ... L2 L2 L3 L3 ... L3 L3 ...
the compressor will compress a little bit less (1 more full list
transmission).

On network streams with lists that follow the pattern:
  L1 L1 ... L1 L1 L2 L1 L1 ... L1 L1 L3 L1 L1 ...
the compressor will compress a little bit more (1 byte saved per
packet).