TCP profile: allow co_common/seq_8/rnd_8 packet types in FO state
According to RFC6846, §5.3.1.4, the decompressor may accept all packets
that are protected by CRC-8 or CRC-7, ie. IR, IR-DYN, co_common, seq_8
and rnd_8. The compressor is thus allowed to use those packet types in
FO state.

This change allows the compressor to use co_common, seq_8 and rnd_8
packet types in addition to the IR and IR-DYN packet types.