decompression: early reject packet types not allowed by context state
Some packet types shall not be accepted by the decompressor depending
on the related context state.

For RFC3095 profiles, the check was performed while packets were parsed.
So, a part of the packet was processed for nothing. And chance are higher
that the decompression context is modified while it should not.

For the TCP profile, the check was not performed except for IR in NC
state.

This change performs the checks in a generic way for all profiles. The
checks are performed earlier before packet is processed by the profile.