Uncompressed profile: add upward mode and state transitions
In the Uncompressed profile, upward mode transitions were not handled at
all, a context remained in U-mode forever. This is wrong. The RFC 3095
§5.10.3 says:
  When feedback is received in any state, the compressor switches to
  Bidirectional mode

In the Uncompressed profile, upward state transitions were handled only
with the optimistic way (ie. a small number of packets is sent to gain
confidence). This is wrong. There is another way to perform an upward
state transition. A positive ACK may cause the transition from IR state
to Normal state too since it gives confidence to the compressor that the
decompressor successfully received at least one IR packet.