Uncompressed profile: IR state required for ambiguous packets
In the Uncompressed profile, some uncompressed packets cannot be
transmitted as Normal packets. The first byte would be mis-interpreted
as another ROHC packet.

An IR packet is therefore required. As IR packets cannot be transmitted
in the Normal state (RFC3095, §5.10.3, b), the compressor shall transit
backward to the IR state (and transmit an IR packet).