add test for fuzzing of the ROHC decompression
The non_regression test is ideal for fuzzing the ROHC compression for
robustness and also the whole compression/decompression process for
end-to-end correctness. However this test does not allow to fuzz the
ROHC decompression for robustness.

So, this change alters the robustness test 'malformed_rohc_packets' to
ease fuzzing the ROHC decompression.

Prepare for bug fixes as soon as AFL will find a problem!