CI: disable warnings about unreachable code
Travis CI runs an outdated version of Clang (3.4) that warns about
unreachable code in some switch statements on enums with an extra
default label. Clang 3.5 works fine.

The default label is required to be robust to users that do not respect
the enums. So, disable the warning for Travis CI.