RFC3095: better handle transition from 1 to 2 IP headers
An assertion occured if a ROHC packet introduced a 2nd IP header w/o
bits for the new IP header. The decompressor tried to retrieve the
information from an uninitialized part of the context.

This change adds checks for context initialization before using the
context information. If not initialized, decompression fails.

Detected with AFL. Several robustness tests were added.