Uncompressed profile: do not create NACK feedback
The decompressor shall never send NACK or STATIC-NACK feedbacks for
the Uncompressed profile. The decompressor was doing so when receiving
a IR-DYN packet with profile ID set to 0x0000 without existing context.

In such a case, the decompressor checked the lack of existing context
before the profile ID, and therefore transmitted a STATIC-NACK feedback.
This was done this way because for most of the packet types the profile
ID is known only if the context is found. This is however not true for
IR and IR-DYN packets.

Problem detected with AFL. A robustness test was added.