decompression: handle NACK for unknown packet types
When encountering an unknown packet type, the decompressor may send
a NACK feedback to the remote compressor. When determining whether
a NACK feedback shall be sent or not, an assertion caused a crash.

This change ensures that the assertion is removed. It is indeed valid to
determine whether an unknown packet contains a CRC-7/CRC-8 or not. Answer
is 'no' of course, but no assertion causes a crash from now.

Detected with AFL. A robustness test was added.