RFC3095 list compression: factorize PS computation
The list compression types 0, 1 and 3 may use 4-bit or 8-bit indexes.
The PS bit determines the length of the indexes:
 * PS = 0 for 4 bits,
 * PS = 1 for 8 bits.

The PS computation depends on the largest index that needs to be
transmitted in the compressed list. The computation is the same for
types 0, 1 and 2. This change factorizes that code in a new dedicated
function.