TCP profile: correctly handle large TS values in TS option
The TCP profile encodes the TS request and reply values of the TS option
with the W-LSB algorithm. Different values of the shift parameter p are
used: -1, 0x40000, 0x4000000.

The ROHC library was using 0x40000 instead of 0x4000000 for large TS
values. This change fixes that behavior to be RFC compliant.

Problem detected with AFL. A non-regression test was added.