TCP profile: support context recycling with changing IPv6 Flow Label
The recycling of a TCP decompression context with another IPv6/TCP
stream with a different IPv6 Flow Label than the previous IPv6/TCP
stream fails. For IR packets the decompressor outputs an IPv6/TCP
packet with the new Flow Label, but for co_common packet the
decompressor outputs an IPv6/TCP packet with the Flow Label of the
1st stream instead of the new Flow Label.

This is due to the decompressor that doesn't record the IPv6 Flow Label
in the decompression context. This change makes the decompressor record
the new value instead of the old one.

Problem detected with AFL. A non-regression test generated by AFL was
added.

The RawPacketIPv6Tunnel-UK6x.cap capture has non-zero IPv6 Flow Label.
It was retrieved from https://wiki.wireshark.org/SampleCaptures and was
also added as non-regression test.