TCP profile: simplify the check of TCP profile
The function c_tcp_check_profile() checks whether a packet could be
compressed with the TCP profile. This function is large. This makes
it difficult to read.

This change splits the verification of the IPv6 extension headers
into a smaller, separate function named:
   rohc_comp_tcp_are_ipv6_exts_acceptable().

This change makes GNU complexity happier: score goes from 73 to 28.