build: warn about functions that require large frames
Update the configure script so that the compilers warn about functions
that require frames larger than 1024 bytes.

Use the -Wframe-larger-than=1024 flag for that purpose. The configure
option --enable-fail-on-warning does not enforce those warnings as
errors for the moment.

The Linux kernel requires smaller function frames than the ROHC library
currently uses. That makes the kernel module fail. The warning will help
fixing the problem.