Reduce function frame size for extensions/options counters
Some functions require to count the number of IPv6 extension headers or
TCP options by types. There is a limit on the number of allowed
occurrences, and that limit shall be checked.

Those extensions and options are counted in an array of size_t. This
leads to large arrays and large function frames.

This change replaces the size_t arrays with uint8_t arrays. This is
possible because the number of expected extensions/options is quite low
(less than 255, probably in the range [0-2]).