build: make warning about functions that require large frames fatal
Previous changes fixed the functions that required large frames. So, the
compiler warning that was added earlier can now be made a fatal error if
--enable-fail-on-warning is specified.

This way, new problems of this kind will be detected before they cause
any harm.

However, do not make such warnings fatal when building with Travis CI.