TCP profile: get rid of unused, duplicated or harmful structs/unions
The following structs were unused:
* ip_rout_opt_t
* ip_hop_opt_t
* ip_dest_opt_t
* ipv6_extension_t

The following structs were duplicates of other existing structs since
the TCP profile was introduced in the ROHC library:
* base_header_ip_v4_t
* base_header_ip_v6_t

The union base_header_ip_t was considered as harmful. It makes the type
verification system useless since there is one type everywhere. There is
no way to check that a function calls another with the expected type.

The struct base_header_ip_vx_t was useful. It was however temporarily
removed during the cleanup. It will be introduced again later.