TCP profile: handle generic_stable_irregular() encoding
The ROHC compressor had no support for the generic_stable_irregular()
encoding, so the decoding part in the ROHC decompressor was untested and
wrong.

This change adds support for the generic_stable_irregular() encoding to
the ROHC compressor. It also fixes the support at decompressor.

The problem was reported and fixed by Klaus Warnke in bug #1495527.