[fix] TCP profile: avoid mixing positions and indexes of TCP options
Every TCP option is identified by a position in the list of TCP options.
This position may change with every new packet. One TCP packet may
contain up to 40 options. The TCP profile supports up to 15 options.

Every TCP option is also identified by an index that allows compressing
it at compressor and decompressing it at decompressor. That index is
chosen in range [0 ; 15].

Some arrays in the volatile or persistent contexts are indexed with the
option position. Some others are indexed with the option index. Some
code paths were mixing the two. This could lead to out-of-bounds reads.
This change fixes the wrong usages that were detected.