TCP profile: fix implementation of scaled ACK
Scaled ACK shall not use the payload size for its encoding. The ROHC
standard proposes to use the deltas of the sequence number of the
related TCP flow running in the opposite direction (if any). Detecting
the related TCP flow is complicated, so this changes estimates the
values with the deltas in the ACK number instead. This change updates
existing tests in consequence.

The changes also:
* fix the rnd_2 packet wrt the PSH flag
* improve packet decision
* improve code coverage

Fix bug #1507675 reported by Klaus Warnke.