configure: gnuplot is not mandatory at buidtime, but at runtime
Do stop the configure script if the stats app is enabled but gnuplot
is not found. gnuplot is not required to build the library or the tool,
only to run the tool.

Turn the error message into a warning message that warn the user that
he/she should install gnuplot, but do not halt the configure script any
more. This way RPM build does not require gnuplot any more.