RFC3095: conformance: add D flag for IP-only, UDP, ESP and UDP-Lite profiles
All RFC3095-based profiles (except the Uncompressed profile) use the x
flag of the IR packet to indicate the presence of the dynamic chain
(ie. they use it as the D flag). The Uncompressed profile shall put 0
in the reserved bit instead.

The ROHC library was not conform to the RFC3095 standard. The x flag of
the IR packet was used as the D flag only for the RFC3095 RTP profile.
So, the IP-only, IP/UDP, IP/ESP and IP/UDP-Lite profiles were wrong.

The problem was reported by Selvaganesan Murugesan at:
  https://answers.launchpad.net/rohc/+question/661412