TCP: avoid assertion with malformed TCP options
If a compressor sends malformed TCP options to the decompressor (length
of TCP options not being a multiple of 32-bit words), the decompressor
shall not assert, but just reject the packet.

Ensure that such malformed packets reports one ROHC_STATUS_MALFORMED
error status to the user.

Also add a new non-regression test to avoid the problem to come back
later.