cleanup: RFC3095: do not record packet type in temporary context
The RFC3095 context contains a temporary part that is only valid during
the compression of one packet. The packet type was recorded in that
temporary context after packet decision, then read during the next steps
of the compression. This change avoids recording the packet type in the
temporary context and transmits the information from functions to
functions with an explicit parameter. This makes the transmission
explicit, therefore making code clearer.

The remaining attributes that are recorded in the temporary context may
eventually completely removed in the future for the very same reason.