TCP: handle length-changing generic options
TCP options that are not well-known to the ROHC protocol are compressed
with the 'generic option' compression scheme. This scheme does allow the
option to change of content by transmitting it in both the dynamic or
irregular chain, but it does allow the option to change of length only
in the dynamic chain.

The compressor did not check the generic options for length changes.
This fix adds the check and forces the compressor to use the dynamic
chain for compressing TCP options. It therefore forces the compressor to
use one IR, IR-DYN, co_common, seq_8 or rnd_8 packet.

An example of generic option is the Multipath TCP option (MTCP). The
wiki of the Wireshark project contains a PCAP capture named
redundant_stream1.pcapng that contains a MTCP stream with at least one
MTCP option that changes of length. This capture was added as new
non-regression test.